Check for file before  trying to remove it

+++ scripts/install_helm.sh
@@ -10,7 +10,11 @@ download_and_install_helm() {
   fi
   local _archive_url=https://get.helm.sh/helm-v${_version//v}-$(get_kernel)-$(get_machine).tar.gz
   local _tmpdir=$(mktemp -d)
-  rm ~/.curlrc
+  
+  if [ -f ~/.curlrc ]; then
+    rm ~/.curlrc
+  fi
+
   curl -o "${_tmpdir}/helm.tar.gz" -fL "${_archive_url}"
   pushd "${_tmpdir}"
     tar -xf helm.tar.gz
